Personal Agent /requirements.txt
python-dotenv
phidata
openai
PyGithub
firecrawl-py
duckduckgo-search